requirements.txt
numpy
pyrtlsdr
bleak>=0.21.0
